Move important modules to top-level (#304)
This PR re-organizes the `utils` subpackage. The important modules are
moved to top-level while internal modules are kept in `utils`.

This is to separate modules that are meant for user use from modules
that are meant for internal use. This way we are able to exclude the
internal modules from the documentation and make using the important
modules easier at the same time.

Old modules are deprecated and will be removed in a future version.
