+++ roseau/load_flow/io/dgs/__init__.py
@@ -33,7 +33,7 @@ from roseau.load_flow.models import (
     VoltageSource,
 )
 from roseau.load_flow.typing import Id, StrPath
-from roseau.load_flow.utils._exceptions import find_stack_level
+from roseau.load_flow.utils import find_stack_level
 
 logger = logging.getLogger(__name__)
