+++ roseau/load_flow/models/core.py
@@ -9,8 +9,7 @@ from shapely.geometry.base import BaseGeometry
 
 from roseau.load_flow.exceptions import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow.typing import Id
-from roseau.load_flow.utils import Identifiable, JsonMixin
-from roseau.load_flow.utils._exceptions import find_stack_level
+from roseau.load_flow.utils import Identifiable, JsonMixin, find_stack_level
 from roseau.load_flow_engine.cy_engine import CyElement
 
 if TYPE_CHECKING:
