+++ roseau/load_flow/utils/__init__.py
@@ -1,67 +1,86 @@
 """
-This module contains utility classes and functions for Roseau Load Flow.
+Internal utility classes and functions for Roseau Load Flow.
 """
 
-from roseau.load_flow.utils.constants import (
-    ALPHA,
-    ALPHA2,
-    DELTA_P,
-    EPSILON_0,
-    EPSILON_R,
-    MU_0,
-    MU_R,
-    OMEGA,
-    PI,
-    RHO,
-    SQRT3,
-    TAN_D,
-    F,
-    NegativeSequence,
-    PositiveSequence,
-    ZeroSequence,
-)
-from roseau.load_flow.utils.mixins import CatalogueMixin, Identifiable, JsonMixin
-from roseau.load_flow.utils.types import (
+from roseau.load_flow.utils.dtypes import (
+    DTYPES,
     BranchTypeDtype,
-    Insulator,
-    LineType,
     LoadTypeDtype,
-    Material,
     PhaseDtype,
     SequenceDtype,
     VoltagePhaseDtype,
 )
+from roseau.load_flow.utils.exceptions import find_stack_level
+from roseau.load_flow.utils.log import set_logging_config
+from roseau.load_flow.utils.mixins import CatalogueMixin, Identifiable, JsonMixin
+from roseau.load_flow.utils.versions import show_versions
 
 __all__ = [
-    # Constants
-    "DELTA_P",
-    "EPSILON_0",
-    "EPSILON_R",
-    "F",
-    "MU_0",
-    "MU_R",
-    "OMEGA",
-    "PI",
-    "RHO",
-    "SQRT3",
-    "TAN_D",
-    "ALPHA",
-    "ALPHA2",
-    "PositiveSequence",
-    "NegativeSequence",
-    "ZeroSequence",
     # Mixins
     "Identifiable",
     "JsonMixin",
     "CatalogueMixin",
-    # Types
-    "LineType",
-    "Material",
-    "Insulator",
-    # Dtypes
-    "PhaseDtype",
-    "VoltagePhaseDtype",
+    # Exceptions and warnings
+    "find_stack_level",
+    # DTypes
+    "DTYPES",
     "BranchTypeDtype",
     "LoadTypeDtype",
+    "PhaseDtype",
     "SequenceDtype",
+    "VoltagePhaseDtype",
+    # Versions
+    "show_versions",
+    # Logging
+    "set_logging_config",
 ]
+
+
+def __getattr__(name: str):
+    import warnings
+
+    deprecation_template = (
+        "Importing {name} from 'roseau.load_flow.utils' is deprecated. Use 'rlf.{module}.{name}' instead."
+    )
+    if name in (
+        "ALPHA",
+        "ALPHA2",
+        "PI",
+        "SQRT3",
+        "DELTA_P",
+        "EPSILON_0",
+        "EPSILON_R",
+        "F",
+        "MU_0",
+        "MU_R",
+        "OMEGA",
+        "RHO",
+        "TAN_D",
+    ):
+        warnings.warn(
+            deprecation_template.format(name=name, module="constants"),
+            category=FutureWarning,
+            stacklevel=find_stack_level(),
+        )
+        from roseau.load_flow import constants
+
+        return getattr(constants, name)
+    elif name in ("PositiveSequence", "NegativeSequence", "ZeroSequence"):
+        warnings.warn(
+            deprecation_template.format(name=name, module="sym"),
+            category=FutureWarning,
+            stacklevel=find_stack_level(),
+        )
+        from roseau.load_flow import sym
+
+        return getattr(sym, name)
+    elif name in ("LineType", "Material", "Insulator"):
+        warnings.warn(
+            deprecation_template.format(name=name, module="types"),
+            category=FutureWarning,
+            stacklevel=find_stack_level(),
+        )
+        from roseau.load_flow import types
+
+        return getattr(types, name)
+    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
